NewThings/requirements.txt
pandas>=1.5
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tensorflow-cpu>=2.10.0
keras>=2.10.0
statsmodels>=0.13.0
scipy>=1.7.0
yfinance>=0.1.70
fredapi>=0.5.0
arch>=5.0.0
joblib>=1.1.0
tqdm>=4.62.0
